WIP Android: rework backend to use game_activity crate
This experimentally updates the Android backend to use the
game_activity create instead of ndk-glue, which enables building
Android applications based on the GameActivity class that's part
of Google's AGDK, instead of NativeActivity.

Incidentally I did also spot several things with the previous
implementation that looked like they may be bugs / issue which I'll
try to file some issues for.