WIP: Windows: deliver a Resumed event after StartCause::Init
To be more consistent with mobile platforms this updates the Windows
backend to dispatch a Resumed event right after the initial
NewEvents(StartCause::Init) event.

It could be beneficial for applications if they could rely on the
Resumed event across all platforms and potentially use that as the place
where they initialize graphics state and render surfaces. This would
potentially set applications up to be more portable by default.