macos: on the main thread, queue redraws via setNeedsDisplay
The aim is that all RedrawRequested callbacks should only ever be driven
through drawRect callbacks.

The next step will be to queue redraws with a reference to the View or
Window so that setNeedsDisplay can also be used in cases where the
request comes from a non-main thread.

TODO: also evaluate how this works with MTLView and OpenGLView, where
it's expected that (by default) setNeedsDisplay is a NOP because
drawRect is anyway called regularly without needing to queue anything.