Add examples/window_pump_events_rfd
The way that the `rfd` crate and Winit interact is poorly defined
and fragile.

Applications are even more likely to break assumptions if they
have an external event loop based on `pump_events`.

This example can be used to smoke test how the use of the `rfd` crate
currently interacts with applications based on an external event loop.